pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pcu20"
version = "0.1.0"
description = "Modern network file server for Siemens Sinumerik PCU20 CNC controllers"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "uvicorn>=0.30",
    "fastapi>=0.115",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "websockets>=13",
    "pydantic>=2.0",
    "tomli-w>=1.0",
    "click>=8.0",
    "structlog>=24.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.5",
]

[project.scripts]
pcu20 = "pcu20.cli:main"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/pcu20"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
